examples/basic: use svelvet ^0.2.0 (#35)
This upgrades the example to use the newest release.

Motivation: I was confused why the basic example was not serving anything on localhost, when according to the README it should have.

Note: npm semver `^` caret ranges fix the left-most non-zero digit. The example will need updating again when you release 0.3.0, which may or may not be what you want.